docker-release: also push images to public repo (#1093)
We've already got everything set up to allow the `docker-release`
workflow to impersonate a GCP SA that is allowed to push images to
`divviup-public-artifacts`. We just need to tag the built image for the
public repo as well as the private one, and then push to both. I also
harmonized the names of the GitHub Actions Environment Variables (TM)
for the workload identity providers and service accounts with the scheme
used in Janus' Docker publishing action.