compose.yaml: autopair first-party leader
To enable the demo workflow, users will need to pair two aggregators
with the control plane. Further, one of them must be first-party, which
can only be done using a CLI built with feature `admin`. But the one we
distribute via GitHub releases isn't, and I argue it shouldn't be, which
means that demo users can't appropriately pair aggregators.

We could provide a `curl` or `wget` command that'd do it, but as it
turns out, we already bundle the `divviup` CLI in the `divviup_api`
container, so let's just use that, since the API URL and token for the
aggregators are static!

I'm thinking we should have the demo user pair the other aggregator
themselves with `divviup` to simulate what they might have to do in a
realistic use case, if they bring their own helper.

Building on previous changes, this commit moves the `pair_aggregator`
service into the main `compose.yaml`, leaving `build` and `develop`
stanzas for it in `compose.dev.override.yaml`. Additionally, the
`divviup_api_integration_test` image version is bumped to one built with
feature `admin`, which should get the tests working.

Part of #1096