Commas are illegal in GHA cache keys (#1101)
In #1100, we started building `divviup_api_integration_test` with
feature `admin`. This breaks `docker-release.yml`, because the Rust
features are used to construct GH Actions cache keys, and commas are
illegal there (at the very least, `docker buildx build --cache-from`
doesn't like it). Construct the cache key by joining with `-` to work
`around this.

Additionally, this _should_ have broken `docker.yml`, but I forgot to
add the `admin` feature there, which this PR corrects.

Part of #1096