Add the "Bring Your Own" functionality
The basic setup is with Docker Compose, and Terraform is added on top
for doing a public cloud setup. Each component is given its own Compose
and Terraform setup. HTTP-only can be used for private setups, and HTTPS
with automatic certificate renewal can be used for public cloud setups.

The code is structured to allow support for multiple cloud providers.
This commit only adds support for Amazon AWS as a starting point.

The directory gets packaged as divviup-bring-your-own.tar.gz and added
to the release artifacts via the existing GitHub Actions workflow. The
Docker image versions for divviup* and janus* sit on :latest in the
repository, but get pinned in the release artifact.

There is also a companion commit for the divviup/public-docs repository
with documentation on how to use divviup-bring-your-own.tar.gz.